@@ -1,42 +1,44 @@
 {
-  "name": "template",
-  "version": "0.0.0",
-  "private": true,
+  "name": "care-pets_back-end",
+  "version": "1.0.0",
+  "description": "",
+  "main": "app.js",
   "scripts": {
     "start": "cross-env NODE_ENV=production node ./server.js",
     "start:dev": "cross-env NODE_ENV=development nodemon ./server.js",
     "lint": "eslint **/*.js",
-    "lint:fix": "eslint --fix **/*.js",
-    "test": "jest",
-    "test:dev": "jest --watch"
+    "lint:fix": "eslint --fix **/*.js"
   },
+  "keywords": [],
+  "author": "'Aleksandr Pavlenko' <aleksandor2138@gmail.com>",
+  "license": "ISC",
   "dependencies": {
+    "@joi/date": "^2.1.0",
     "bcrypt": "^5.1.0",
-    "cors": "2.8.5",
-    "cross-env": "7.0.3",
+    "bcryptjs": "^2.4.3",
+    "body-parser": "^1.20.1",
+    "cloudinary": "^1.34.0",
+    "cors": "^2.8.5",
+    "cross-env": "^7.0.3",
     "dotenv": "^16.0.3",
-    "express": "4.17.1",
-    "gravatar": "^1.8.2",
+    "express": "^4.18.2",
     "http-errors": "^2.0.0",
-    "jimp": "^0.16.2",
-    "joi": "^17.7.0",
+    "jimp": "^0.22.4",
+    "joi": "^17.7.1",
     "jsonwebtoken": "^9.0.0",
-    "mongoose": "^6.9.0",
-    "morgan": "1.10.0",
+    "mongoose": "^6.9.1",
+    "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
-    "nanoid": "^3.3.4",
-    "nodemailer": "^6.9.1"
+    "swagger-ui-express": "^4.6.0",
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "@types/jest": "^29.4.0",
-    "eslint": "7.19.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-standard": "16.0.3",
-    "eslint-plugin-import": "2.25.3",
+    "eslint": "^8.34.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "5.2.0",
-    "jest": "^29.4.1",
-    "nodemon": "2.0.15",
-    "supertest": "^6.3.3"
+    "eslint-plugin-promise": "^6.1.1",
+    "nodemon": "^2.0.20"
   }
 }
