@@ -1,51 +1,51 @@
-import React, { lazy, useEffect } from 'react';
-import { useDispatch} from 'react-redux';
-// import { fetchAll } from 'redux/contacts/requestAPI';
-// import { selectIsLoading, selectError } from 'redux/contacts/contactsSlise';
-import { Route, Routes } from 'react-router-dom';
-import { HeaderAll } from './Header/Header';
-import { fetchCurrentUser } from 'redux/authorization/requestAPI';
-import {PrivateRoute} from './PrivateRouter/PrivateRouter';
-import { PublicRoute } from './PublicRouter/PublicRouter';
-import { useAuth } from 'hooks';
+import { Routes, Route, Navigate } from 'react-router-dom';
+import { lazy, Suspense} from 'react';
+import SharedLayout from 'components/SharedLayout';
+import PrivateRoutes from './PrivateRoutes';
+import PublicRoutes from './PublicRotes';
+import { getToken } from 'redux/selectors';
+import { useGetCurrentUserQuery } from 'redux/fetchUser';
+import NoInternetConnection from './NointernetConnection/NoInternetConnection';
+import { useSelector } from 'react-redux';
+const Home = lazy(() => import('pages/Home'));
+const NewsPage = lazy(() => import('pages/NewsPage'));
+const OurFriendsPage = lazy(() => import('pages/OurFriendsPage'));
+const NoticesPage = lazy(() => import('pages/NoticesPage'));
+const LoginPage = lazy(() => import('pages/LoginPage'));
+const RegisterPage = lazy(() => import('pages/RegisterPage'));
+const UserPage = lazy(() => import('pages/UserPage'));
+const NoticesCategoriesList = lazy(() =>
+  import('components/NoticesCategoriesList')
+);
 
-const PageNotFound = lazy(() => import('../pages/PageNotFound/PageNotFound'));
-const HomePage = lazy(() => import('../pages/home'));
-const SignUp = lazy(() => import('../pages/SignUp'));
-const SignIn = lazy(() => import('../pages/SignIn'));
-const PhoneBook = lazy(() => import('../pages/Phonebook'));
+export const App = () => {
 
-const App = () => {
-  const dispatch = useDispatch();
-  const { isRefreshing } = useAuth();
+  const token = useSelector (getToken);
+  useGetCurrentUserQuery({skip : !token})
 
-  useEffect(() => {
-    dispatch(fetchCurrentUser());
-  }, [dispatch]);
-
-  return isRefreshing ? (
-    <b>Refreshing user...</b>
-  ) : (
-    <Routes>
-      <Route path="/" exact element={<HeaderAll />}>
-        <Route index element={<HomePage />} />
-        <Route
-          path="phonebook"
-          element={
-            <PrivateRoute redirectTo="/sign-in" component={<PhoneBook />} />
-          }
-        />
-        <Route
-          path="sign-up"
-          element={<PublicRoute redirectTo="/phonebook" component={<SignUp />} />}
-        />
-        <Route
-          path="sign-in"
-          element={<PublicRoute redirectTo="/phonebook" component={<SignIn />} />}
-        />
-        <Route path="*" element={<PageNotFound />} />
-      </Route>
-    </Routes>
+  return (
+    <NoInternetConnection>
+      <Suspense>
+        <Routes>
+          <Route path="/" element={<SharedLayout />}>
+            <Route index element={<Home />} />
+            <Route path="/news" element={<NewsPage />} />
+            <Route path="/notices" element={<NoticesPage />}>
+              <Route index element={<Navigate to="sell" />}></Route>
+              <Route path=":category" element={<NoticesCategoriesList />} />
+            </Route>
+            <Route path="/friends" element={<OurFriendsPage />} />
+            <Route element={<PublicRoutes restricted />}>
+              <Route path="/login" element={<LoginPage />} />
+              <Route path="/register" element={<RegisterPage />} />
+            </Route>
+            <Route element={<PrivateRoutes />}>
+              <Route path="/user" element={<UserPage />} />
+            </Route>
+          </Route>
+          <Route path="*" element={<Navigate to="/" />} />
+        </Routes>
+      </Suspense>
+    </NoInternetConnection>
   );
 };
-export default App;
