@@ -1,22 +1,23 @@
-require("dotenv").config();
-const mongoose = require("mongoose");
-mongoose.set("strictQuery", true);
-mongoose.set("debug", true);
+const { mongoose } = require("mongoose");
+
+const dotenv = require("dotenv");
+dotenv.config();
+
 const app = require("./app");
 
-const PORT = process.env;
-const HOST_URI = process.env;
+const { PORT = 3000, MONGO_URI } = process.env;
 
-(async function () {
-    try {
-        await mongoose.connect(HOST_URI);
-        console.log("Database connection successful");
+mongoose.Promise = global.Promise;
+mongoose.set("strictQuery", true);
 
-        app.listen(PORT, () => {
-        console.log(`Server running. Use our API on port: ${PORT}`);
-        });
-    } catch (error) {
-        console.error("Error while connecting to mongodb", error.message);
-        process.exit(1);
-    }
-})();
+try {
+  mongoose.connect(MONGO_URI);
+  console.log("Database connected succsesful");
+
+  app.listen(PORT, () => {
+    console.log("Server running. Use our API on port:", PORT);
+  });
+} catch (error) {
+  console.log(error.message);
+  process.exit(1);
+}
